Feature/bevy tiles 2d (#3)
* Split 2d and 3d

+++ crates/bevy_tiles/src/commands/tile_batch.rs
@@ -8,9 +8,9 @@ use super::{insert_tile_batch, take_tile_batch};
 
 pub struct SpawnTileBatch<F, B, IC, const N: usize = 2>
 where
-    F: Fn([isize; N]) -> B + Send + 'static,
+    F: Fn([i32; N]) -> B + Send + 'static,
     B: Bundle + Send + 'static,
-    IC: IntoIterator<Item = [isize; N]> + Send + 'static,
+    IC: IntoIterator<Item = [i32; N]> + Send + 'static,
 {
     pub map_id: Entity,
     pub tile_cs: IC,
@@ -19,12 +19,12 @@ where
 
 impl<F, B, IC, const N: usize> Command for SpawnTileBatch<F, B, IC, N>
 where
-    F: Fn([isize; N]) -> B + Send + 'static,
+    F: Fn([i32; N]) -> B + Send + 'static,
     B: Bundle + Send + 'static,
-    IC: IntoIterator<Item = [isize; N]> + Send + 'static,
+    IC: IntoIterator<Item = [i32; N]> + Send + 'static,
 {
     fn apply(self, world: &mut World) {
-        let (tile_cs, bundles): (Vec<[isize; N]>, Vec<B>) = self
+        let (tile_cs, bundles): (Vec<[i32; N]>, Vec<B>) = self
             .tile_cs
             .into_iter()
             .map(|coord| (coord, (self.bundle_f)(coord)))
@@ -33,7 +33,7 @@ where
         let tiles = tile_cs
             .into_iter()
             .zip(world.spawn_batch(bundles))
-            .collect::<Vec<([isize; N], Entity)>>();
+            .collect::<Vec<([i32; N], Entity)>>();
 
         insert_tile_batch::<N>(world, self.map_id, tiles);
     }
@@ -41,7 +41,7 @@ where
 
 pub struct DespawnTileBatch<IC, const N: usize = 2>
 where
-    IC: IntoIterator<Item = [isize; N]> + Send + 'static,
+    IC: IntoIterator<Item = [i32; N]> + Send + 'static,
 {
     pub map_id: Entity,
     pub tile_cs: IC,
@@ -49,7 +49,7 @@ where
 
 impl<IC, const N: usize> Command for DespawnTileBatch<IC, N>
 where
-    IC: IntoIterator<Item = [isize; N]> + Send + 'static,
+    IC: IntoIterator<Item = [i32; N]> + Send + 'static,
 {
     fn apply(self, world: &mut World) {
         for (_, tile_id) in take_tile_batch::<N>(world, self.map_id, self.tile_cs) {
@@ -60,7 +60,7 @@ where
 
 pub struct MoveTileBatch<IC, const N: usize = 2>
 where
-    IC: IntoIterator<Item = ([isize; N], [isize; N])> + Send + 'static,
+    IC: IntoIterator<Item = ([i32; N], [i32; N])> + Send + 'static,
 {
     pub map_id: Entity,
     pub tile_cs: IC,
@@ -68,7 +68,7 @@ where
 
 impl<IC, const N: usize> Command for MoveTileBatch<IC, N>
 where
-    IC: IntoIterator<Item = ([isize; N], [isize; N])> + Send + 'static,
+    IC: IntoIterator<Item = ([i32; N], [i32; N])> + Send + 'static,
 {
     fn apply(self, world: &mut World) {
         const ERR_MESSAGE: &str =
@@ -77,12 +77,12 @@ where
         let mut tile_cs = self
             .tile_cs
             .into_iter()
-            .collect::<HashMap<[isize; N], [isize; N]>>();
+            .collect::<HashMap<[i32; N], [i32; N]>>();
 
         let removed = take_tile_batch::<N>(
             world,
             self.map_id,
-            tile_cs.keys().cloned().collect::<Vec<[isize; N]>>(),
+            tile_cs.keys().cloned().collect::<Vec<[i32; N]>>(),
         )
         .into_iter()
         .map(|(tile_c, tile_id)| (tile_cs.remove(&tile_c).expect(ERR_MESSAGE), tile_id));
@@ -93,7 +93,7 @@ where
 
 pub struct SwapTileBatch<IC, const N: usize = 2>
 where
-    IC: IntoIterator<Item = ([isize; N], [isize; N])> + Send + 'static,
+    IC: IntoIterator<Item = ([i32; N], [i32; N])> + Send + 'static,
 {
     pub map_id: Entity,
     pub tile_cs: IC,
@@ -101,7 +101,7 @@ where
 
 impl<IC, const N: usize> Command for SwapTileBatch<IC, N>
 where
-    IC: IntoIterator<Item = ([isize; N], [isize; N])> + Send + 'static,
+    IC: IntoIterator<Item = ([i32; N], [i32; N])> + Send + 'static,
 {
     fn apply(self, world: &mut World) {
         const ERR_MESSAGE: &str =
@@ -110,12 +110,12 @@ where
         let tile_cs = self
             .tile_cs
             .into_iter()
-            .collect::<BiMap<[isize; N], [isize; N]>>();
+            .collect::<BiMap<[i32; N], [i32; N]>>();
 
         let removed_left = take_tile_batch::<N>(
             world,
             self.map_id,
-            tile_cs.left_values().cloned().collect::<Vec<[isize; N]>>(),
+            tile_cs.left_values().cloned().collect::<Vec<[i32; N]>>(),
         )
         .into_iter()
         .map(|(tile_c, tile_id)| (*tile_cs.get_by_left(&tile_c).expect(ERR_MESSAGE), tile_id));
@@ -123,7 +123,7 @@ where
         let removed_right = take_tile_batch::<N>(
             world,
             self.map_id,
-            tile_cs.right_values().cloned().collect::<Vec<[isize; N]>>(),
+            tile_cs.right_values().cloned().collect::<Vec<[i32; N]>>(),
         )
         .into_iter()
         .map(|(tile_c, tile_id)| (*tile_cs.get_by_right(&tile_c).expect(ERR_MESSAGE), tile_id));
